+++ src/main/java/com/example/demo/staff/Staff.java
@@ -14,7 +14,7 @@ import lombok.*;
 public class Staff {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long staffId;
+    private Integer staffId;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "userId")
